% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hycom-package.R
\docType{package}
\name{hycom}
\alias{hycom}
\title{hycom: Simplified access to HYCOM (HYbrid Coordinate Ocean Model) data from R.}
\description{
Simplified access to HYCOM (HYbrid Coordinate Ocean Model) data from R.
}
