% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncdf.R
\name{hycom_depth}
\alias{hycom_depth}
\title{Retrieve the HYCOM depth vector}
\usage{
hycom_depth(x)
}
\arguments{
\item{x}{ncdf4 object}
}
\value{
a vector of depths
}
\description{
Retrieve the HYCOM depth vector
}
