% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncdf.R
\name{hycom_most_recent_time}
\alias{hycom_most_recent_time}
\title{Retrieve the most recent timestep}
\usage{
hycom_most_recent_time(x)
}
\arguments{
\item{x}{ncdf4 object}
}
\value{
POSIXct time
}
\description{
Retrieve the most recent timestep
}
