% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncdf.R
\name{hycom_res}
\alias{hycom_res}
\title{Retrieve HYCOM spatial resolution}
\usage{
hycom_res(x)
}
\arguments{
\item{x}{ncdf4 object}
}
\value{
two element [x, y] resolution
}
\description{
Retrieve HYCOM spatial resolution
}
\seealso{
[HYCOM](https://www.hycom.org/dataserver/gofs-3pt1/analysis)
}
