% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncdf.R
\name{hycom_res_lut}
\alias{hycom_res_lut}
\title{Retrieve HYCOM spatial resolution as a per element look up table}
\usage{
hycom_res_lut(x)
}
\arguments{
\item{x}{ncdf4 object}
}
\value{
two element [x, y] list of resolution vectors
}
\description{
Retrieve HYCOM spatial resolution as a per element look up table
}
\seealso{
[HYCOM](https://www.hycom.org/dataserver/gofs-3pt1/analysis)
}
