% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncdf.R
\name{hycom_t0}
\alias{hycom_t0}
\title{Retrieve the HYCOM time origin}
\usage{
hycom_t0(x, tz = "UTC")
}
\arguments{
\item{x}{ncdf4 object}

\item{tz}{character, the time zone to use}
}
\value{
POSIXct time
}
\description{
Retrieve the HYCOM time origin
}
