% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hycom.R
\name{hycom_url}
\alias{hycom_url}
\title{Craft a HYCOM URL for a given date}
\usage{
hycom_url(
  filename = c("ts3z", "uv3z", "ssh")[1],
  product = "GLBy0.08",
  version = "expt_93.0",
  root = file.path("http://tds.hycom.org/thredds/dodsC")
)
}
\arguments{
\item{filename}{character, the name of the file. Filenames are
coded to reflect the variables served.}

\item{product}{character the product to fetch}

\item{version}{character the version to fetch}

\item{root}{character, the root URL}
}
\value{
one or more URLs
}
\description{
Craft a HYCOM URL for a given date
}
\seealso{
the \href{https://tds.hycom.org/thredds/catalog.html}{HYCOM THREDDS catalog}
}
