% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncdf.R
\name{hycom_vardims}
\alias{hycom_vardims}
\title{Retrieve a list of the dimensions for each variable}
\usage{
hycom_vardims(x)
}
\arguments{
\item{x}{ncdf4 object}
}
\value{
named list, one element per variable, identifying the dimensions
}
\description{
Retrieve a list of the dimensions for each variable
}
