% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncdf.R
\name{hycom_varndims}
\alias{hycom_varndims}
\title{Retrieve a list of the number of dimensions for each variable}
\usage{
hycom_varndims(x)
}
\arguments{
\item{x}{ncdf4 object}
}
\value{
named vector, one element per variable, identifying the number of dimensions
}
\description{
Retrieve a list of the number of dimensions for each variable
}
