% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncdf.R
\name{hycom_vars}
\alias{hycom_vars}
\title{Retrieve the names of variables}
\usage{
hycom_vars(
  x,
  drop = c("salinity_bottom", "water_temp_bottom", "tau", "water_u_bottom",
    "water_v_bottom")
)
}
\arguments{
\item{x}{ncdf4 object}

\item{drop}{character, the variable to drop, set to 'none' to drop none.}
}
\value{
character vector of variable names
}
\description{
Retrieve the names of variables
}
